Convert Danishbits to use their api (#2832)
* Delete danishbits.py
* Create danishbits.py
* Update __init__.py
* Tweaks
* Add pubdate
* Special thanks to DanishBits.org
