+++ medusa/providers/generic_provider.py
@@ -478,7 +478,7 @@ class GenericProvider(object):
         """Login to provider."""
         return True
 
-    def search(self, search_params, age=0, ep_obj=None):
+    def search(self, search_strings, age=0, ep_obj=None):
         """Search the provider."""
         return []
