+++ readme.md
@@ -69,6 +69,7 @@ A full list can be found here: [Link](https://github.com/pymedusa/Medusa/wiki/Me
 ![image](https://nzb.cat/favicon.ico)[NZB.cat](https://nzb.cat/)
 ![image](https://nzbgeek.info/favicon.ico)[NZBGeek](https://nzbgeek.info)
 ![image](https://raw.githubusercontent.com/pymedusa/Medusa/master/static/images/providers/dognzb.png)[DOGnzb](https://dognzb.cr)
+![image](https://raw.githubusercontent.com/pymedusa/Medusa/master/static/images/providers/danishbits.png)[DanishBits](https://danishbits.org)
 
 #### Browsers support <sub><sub>made by <a href="https://godban.github.io">godban</a></sub></sub>
